Implemented vm_lock_page() to lock a page into memory.
Locking a page means that it will be kept mapped in and guaranteed to be
safe to access until it is unlocked with vm_unlock_page(). At the moment
the implementation doesn't need to do much because once a page gets
mapped in it will not be unmapped again until the region is unmapped (the
code doesn't actually prevent a region with locked pages from being unmapped
yet, there's a TODO in the code for it). It just ensures that the page is
mapped for the requested access. Once it is possible for pages to be
evicted (e.g. swap, evicting cache pages) then the implementation will get
quite a bit more complicated.

As a side effect of implementing this, fixed a possible SMP bug where if
two threads simultaneously fault on a not present page in an anonymous
region, one will succeed in mapping and the second will crash due to the
mapping already existing.