Fixes to toolchain configuration.
Rather than creating a wrapper script for clang that passes a target
and sysroot option, configure the sysroot path at build time, and
symlink <target>-clang to the main clang binary. From this clang
determines the target name automatically. This also fixes an issue
where clang would attempt to use the system as/ld binaries if the
toolchain is not in PATH.