Cleaned up and improved build system.
Cleaned up the build system a bit and improved the way userspace is built.
Now rather than using a single environment for the whole of userspace, each
app/library creates its own using the new BuildManager class. This class
maintains information about all libraries and their header paths. When
creating a new environment, a list of libraries to be used can be specified,
which sets up the environment with the variables needed to use those
libraries.

This functionality used to be implemented by the KiwiLibrary/KiwiApplication
builders in the userspace environment. This was problematic in cases where
an application/library needed to be built in multiple parts (e.g. with
different flags for different source files): the sub-parts would not have
the header paths for the needed libraries set, so they would have to be
specified manually. Now, everything built in an environment will have header
paths, etc., set up correctly.