Initial implementation of security tokens.
This commit implements security tokens. A security token is a kernel
object containing a security context which is used for security checks
in a process/thread.

Each process has a security token associated with it. In addition, a
thread can set its own security token which overrides the process token,
in order to temporarily take on a different identity and set of
privileges.

A thread cannot create a token object containing privileges that it
does not have, and it cannot create one with a different identity unless
it has the PRIV_CHANGE_IDENTITY privilege.

Handles to token objects will be able to be transferred over IPC in
order to grant privileges to another process. This will be used by the
security server in order to implement user login.