Unset the old thread's user stack in kern_process_exec()
When a thread terminates, if the kernel had created its user stack it will
be automatically unmapped. A new thread is created by an exec, so when the
old thread exited, if it had a user stack pointer set that range would be
unmapped from the new address space when the reaper cleans it up.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>