Kill off object_t
Now that object security has gone, the only purpose that object_t served
was to hold an object_type_t pointer. This adds the overhead of an extra
pointer to each object. Instead we replace the object_t pointer in each
handle with an object_type_t pointer. Object type code now uses the handle
private pointer to point straight to the object, or if it needs extra
per-handle state, to an additional structure (e.g. file_handle_t) which
includes a pointer to the actual object.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>