Add optional allocation tracing output to slab for finding memory leaks
Add a Kconfig option CONFIG_SLAB_TRACING which prints out a message on the
debug console on every allocation and free giving the address/function at
which the allocation/free call took place, and a script which takes a
kernel log with this output and uses it to create a list of all allocations
made which have not been freed by the time the log finishes. This can be
used to find memory leaks.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>