Remove VM_MAP_INHERIT
It was potentially going to be used to pass information to a new process
when created but this is a bad idea since a mapping has the potential to
conflict with where an executable might want to be loaded to. A better
solution is to pass an IPC connection to the child and pass data through
it.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>