Add a putc_unsafe method to console implementations for KDB/fatal to use
This does not take any locks and is intended solely for use within fatal()
calls and KDB where it is guaranteed that other CPUs will be idle.

I hit a deadlock entering KDB because the NMI to trap other CPUs was
received on another core while it held the FB console lock, which caused
KDB to hang.