Remove -Wold-style-cast
While I would like this to be an error for C++ code, clang does not properly
treat our system headers as actual system headers and disable certain warnings
for them. This means that C++ code which uses any C headers that have C-style
casts in get errors if we have this flag.