Start work on terminal infrastructure
This will provide POSIX-style pseudo terminals, implemented as a user mode
service (using a planned kernel interface to allow implementation of file
objects in user mode).

The terminal implementation will live in user mode rather than a kernel module
providing devices, because for a full POSIX terminal implementation, there are
a bunch of POSIXisms that the terminal will need to be aware of (signals,
process groups, etc.). These are unknown to the kernel and will also be
implemented in user mode on top of lower level kernel features.