Basic implementation of the terminal client app
Untested so far since the terminal service itself is unimplemented.

For now this will just relay I/O from the kernel console device through the
terminal service as an initial test for that.

Once we have drivers, then this will use the display/input devices directly,
and later once we get a GUI it will become a GUI application.