Support for creating kernel-to-user IPC connections
Unlike what was there before (removed in the previous commit), we now create a
connection directly without a port, rather than the kernel having to create a
port which the user process then has to connect to.

There is also support for the kernel side to set a callback function for when
the user process sends a message. This allows messages to be processed
directly inside the kern_connection_send() call, rather than having to create
a kernel thread to listen for messages on the connection.