Add user file API
The user file API allows creation of file object handles where operations on
them are implemented by a user mode process (the one which created the file).

This is done via an IPC connection: every operation on a user file results in
a message being sent to the file's creator. The operation will wait until a
reply to that operation is sent back over the connection.

Currently, this just supports basic operations:
 - kern_file_read()
 - kern_file_write()
 - kern_file_info()

No memory mapping or object event support yet. First use case is for the user
mode terminal service.