Handle cancellation of user file operations
If the wait for a user file operation to complete is interrupted, then by the
time the user file's owner replies to the operation, it's serial number will
not be able to be found. This was being treated as an error because I didn't
consider interruption.

Fix this to return STATUS_CANCELLED out of the kern_connection_send() call for
the reply, which the caller can handle however it needs to.