Rework how module state and reference count are managed
They are now stored together in a single atomic variable. This means that we
can atomically try to add a reference to a module with a check that it is not
currently being unloaded.

Also added module_from_addr() to get a module by address, and module_self() to
get the module which calls it.

These will be used to tie devices to modules, so that we can prevent modules
from being unloaded while any device that is maintained by them are in use.