Add initial device resource management functions
This is a facility for allocating resources for devices such that they will be
automatically released when the device they are registered to is destroyed.
This reduces cleanup work that drivers will need to perform when being
unloaded or when hotplugging devices. This is pretty much the same idea as
Linux's devm_* family of functions.