IRQ handling improvements
Main change is that threaded handlers are now the primary recommended way of
handling IRQs, since they are much less restricted in what they can do.

To avoid the need for a handler that runs directly in interrupt context when
the IRQ is received to clear whatever condition caused the IRQ to fire (and
prevent the IRQ from immediately refiring), we now disable the IRQ until the
threaded handler completes.

Other than that, improved the implementation, and fixed various thread safety
issues.