Add a device for basic access to the kernel console framebuffer
This gives userspace access to the framebuffer that is currently being used
for the kernel console. Since we have no mode setting drivers, this is just
whatever framebuffer/mode is set up by KBoot.

This will be initially used for display access from userspace, before we have
a proper display driver framework in future.