Implement destruction of other threads in kern_process_exec()
The code that was checking for other threads was subject to failure due to a
race if a process exec's and then exec's again almost immediately, since on
the second exec, the old thread left over from before the first exec might not
have exited yet.

May as well just implement this properly (however there is a caveat, see the
TODO comment).