Only allow one instance of an object callback interrupt to be pending at a time
Previously each time an event which has a callback is signalled, an interrupt
would be queued to the thread. If the thread currently has its IPL set to
block the interrupt, and the event keeps firing, many interrupts would be
queued to the thread. This could easily be exploited to pile up a large number
of queued interrupts that would never be processed.

Therefore we now drop any events signalled while a previously queued interrupt
is still pending.

This also means we only need to allocate one thread_interrupt_t per callback
and can do this up-front, rather than allocating in object_event_signal().
This has allowed object_event_signal() to be made safe to use in interrupt
context (which was the primary motivation for looking into all this in the
first place).