Remove DPC system
All users of this have now been removed. Threaded IRQ/timer handlers should be
used in its place.

It was an inherently unreliable system, as its whole purpose was to be usable
from interrupt context, but had to deal with an arbitrary number of pending
requests at any one time, which really requires some dynamic allocation. It
preallocated a block of request structures up front but it would be possible
under heavy load to exhaust the supply, and if there's none available at a
call to dpc_request() then there's nothing we can do.