Fix some issues with failure handling in kern_user_file_create()
file_handle_attach() would close the handle on failure resulting in a double
free since we were freeing the user_file on the failure path. Replace
file_handle_attach() with a new file_handle_open() which behaves like
object_handle_open() - does not call the close method on failure.

Also have to be careful since the connection handle is visible to the user
process before the file handle is created, and it is theoretically possible
for another thread to close the connection before we go down the file handle
creation failure path and result in double frees on the IPC connection.