Start defining the network device class interface
Implementation/API for this is a little bit different to existing device
classes/buses, existing code will be updated to be consistent some other time.
I have some planned changes such as separating device tree node creation and
publishing (making the device available to use), this is closer to how things
will be after that.