Remove the platform build configuration
The design of separating architecture and platform, and compiling for a
specific platform, came from a time before the ARM world standardized around
device trees. It assumed that the kernel (when ARM support was eventually
added) would need to have a different version compiled for each supported
platform.

Now that device trees are a thing, this no longer makes sense: we can instead
build a single system that will support all desired platforms.

Therefore, remove the platform from the build configuration. Later on we'll
add back some configuration to allow choosing supported platforms (which will
affected which platform-specific drivers are included), but this will all be
built into the same kernel/system.