Expand initcall mechanism to allow different types of function called at different times
Previously this was just used for functions called late in initialization on
the initialization thread.

This change will allow us to use it for initializing core platform devices
such as IRQ controllers and timer devices which need to be set up earlier. The
set of functions to call for these is dependent on build configuration, and we
don't want to have to hardcode a list of functions.