Require devices to be explicitly published before they can be used
Previously, as soon as device_create() finishes, the created device could be
opened. However device node creation typically needs to be done early in the
process of adding a device in drivers. This gives the possibility of something
trying to open a device before the driver has fully initialised it, which
would lead to problems.

To resolve this we now have separate device creation and publishing steps.
Before a device is published, it (and any child devices) cannot be opened.