Don't use serial port if it's not present
If KBoot doesn't pass a serial tag (because the port's not present), we'd try
to set up a default serial port on PC even if it's not present. This would
result in KDB hanging trying to read input from a port that doesn't exist.

Check that the port exists in the same way as KBoot to fix this.