Some initial work towards being able to create sockets
Since the network stack is in a module this needs some plumbing to be able
to be able to route socket syscalls into the module. This is done by allowing
address families to be registered with an ops structure with the core kernel
socket API, and all operations on sockets with that family are directed
through that ops structure.

This would also allow other socket types such as UNIX sockets to be
implemented in a different module later.