Allow using io_request_{copy,map}() without updating transferred count
These would always update the count, but in cases where we have steps that
could fail after we've copied data out of the request, we had to make sure
that we revert the transferred count on failure. It's cleaner and less error
prone for those cases if we can tell the functions to not advance the count,
and then update it later manually only once we've succeeded.