Add a global rwlock to synchronise access to network address configuration
See comment on net_addr_read_lock(). I'm a bit dubious about adding a global
lock for this and this may turn out to be a bad idea later, but it seems OK as
a simple solution for now, and being a rwlock means it should be too much
overhead (though the rwlock implementation could do with some improvement...).