Require user file servers to specify the operations they support
This allows the kernel to return error for any operations that are unsupported
without sending anything to the server. This is useful for if we were to add
new operations in future without having to worry about whether existing server
code will break if it is sent the new operation.