Refactor how routing information is stored and passed around
We also need a gateway IP to handle routes where we need to send a packet to
a host that is not its final destination. With this as well we have quite a
bit of state to pass around, so bundle it all up into the new net_route_t.