Implement a configurable routing table for IPv4
This adds a routing table for IPv4 which can have entries added and removed
through a device request interface. This includes the ability to add a default
route via a gateway, which we were lacking before.

Routing queries for IPv4 no longer have to search through the network
interface list, we just check the routing table.

As part of this, we now clean up state when an interface has an address
removed or is brought down, including removing ARP cache entries.