Implement kernel process groups API
See description in process_group.h. This is not POSIX process groups: this is
a very basic API provided by the kernel for creating arbitrary groups of
processes, which can automatically track child processes as well.

POSIX process groups and sessions will be built on top of this by userspace
services and APIs.