Start of a C++ utility library
As we add more C++ code to userspace its becoming clear that we could do with
a common utility library. First thing here is a RAII kernel handle wrapper.

This may not be the final structure of this library - I think in the long term
I would like to port Qt for UI so we may need to make changes to better
integrate with that.