Initial implementation of POSIX signals
This is built using the kernel asynchronous object event callback mechanism to
interrupt processes and execute signal handlers. State for signal handling is
managed across libsystem and the POSIX service. Sending signals is done
through a request to the POSIX service, which will deliver the signal to the
process according to its signal handling state.

Processes are only known to the POSIX service if they have made changes from
the default signal handling state. If a signal is directed at an unknown
process, it is treated as just having default signal handling state, and the
signal is delivered using the standard kernel process APIs (i.e.
kern_process_kill()).

Still a bunch of TODOs here, in particular signal masks are not implemented,
and this needs to be hooked up to kernel exception handling so that signal
handlers for exception signals get called in response to exceptions.