Use kernel exceptions to implement abort()
For processes that haven't used POSIX signal APIs we shouldn't unnecessarily
start interacting with the POSIX service sending signals just to abort.
Instead use the kernel exception APIs directly, with which we can also
correctly implement the POSIX-specified behaviour (this will end up calling
through to the SIGABRT handler if one is registered).