Rework thread sleeping/synchronisation to fix race conditions
acquire_wait_lock() was used when waking a thread to acquire the lock for the
synchronisation primitive the thread is waiting on, so that it can be removed
from the waiting threads list.

The implementation of this was racy: to maintain correct locking order to
prevent deadlock, the wait lock needed to be taken before the thread lock.
However, a new wait lock can be set in between acquire_wait_lock() checking
for it and locking it. If that happens the wait lock would not be taken and
we'd hit an assert.

To resolve this, move the responsibility of removing threads from the wait
lists to sync primitive implementations. They have to do this both when waking
a thread off the list, and after waking from sleep in case the sleep was
interrupted.