Fix how rwlock writers are identified
Using a thread flag is racy because thread flags are not atomic and the thread
lock was not taken to set the flag. Also, somehow I'd managed to define
THREAD_RWLOCK_WRITER and THREAD_PREEMPTED to the same value.

Instead use a local waiter structure rather than putting threads directly on
the list, and flag whether the waiter is a writer in there.