Bail out of vm_fault() early if in interrupt context
If we're in interrupt context here it means that a kernel exception has
happened in an interrupt handler or while holding a spinlock. This cannot be
handled, so just bail out. This means we crash with the exception details
rather than hitting the !in_interrupt() assert when locking the address space
mutex.