Replace KDB breakpoint on usermode exceptions with a backtrace to the log
The main reason I had the KDB breakpoint was to get a backtrace for the
exception. This is a bit intrusive though, so replace it with a backtrace of
the thread to the log before we kill it. This uses safe user memory access
functions to access the stack frame/symbols and is limited to 8 frames in
case the user thread has a bad stack.