Implement signal state inheritance across execve()
We need to inherit signal mask, ignored signals and pending signals across an
execve() call. Both the POSIX service and new executable need to have the
correct state after the exceve().

On the service side, some changes are made to treat connection hangup and then
reconnection by the same process as being an execve(), which will reset state
necessary.

In libsystem, state inheritance is implemented through the environment passed
to the new process. We modify the environment passed to execve() to add state
information, which gets picked up (and removed from the environment) during
init in the new executable.