Guarantee that kernel APIs do not return valid but closed handle IDs upon failure
Kernel APIs that return handles do so via a handle_t pointer. This change
makes the guarantee that if these APIs fail, we do not leave a valid but
closed handle ID in the handle return location: APIs will either not write
to the return location, or explicitly write INVALID_HANDLE to it.

There were some cases where functions could fail after writing a handle ID,
in which case they would close the handle but leave the ID. Fix these to
overwrite the value with INVALID_HANDLE.

This guarantee is important for Kiwi::Core::Handle::attach(), which is used
for capturing handles returned by kernel functions into a Handle. If a valid
but closed ID is left in the return location on failure, the Handle would
incorrectly behave as though it owns that ID and try to close it later (by
which time the ID may have been reused by another successful call).