Implement a way to access devices through the filesystem API
This provides a way to access devices for non-native applications. Paths that
start with "//@/" are treated as device paths (with the prefix removed),
supported by kern_fs_open() and kern_fs_info(). They are also supported as
symlink destinations, so you can create links in the regular filesystem to
devices.