Make thread flags atomic, move in_usermem flag into there
thread_t::in_usermem was a standalone bool, but this is wasteful. Previously,
thread_t::flags could only be modified under lock because it was not atomic,
and taking the lock just to manipulate a flag is also wasteful. So, make the
flags atomic (with accessor functions added), and move in_usermem there.