Transfer data inline in ipc_message_t arguments for small user_file I/O ops
There are unused arguments in the user_file messages for I/O operations. We
can make use of this space to store the transfer data if it will fit there
(16 bytes for reads, 8 bytes for writes). This avoids the need to allocate
and send a separate data buffer with the operation. This should improve
performance for small transfers.