Add SO_BINDTOINTERFACE socket option to bind to a specific interface
This is like Linux's SO_BINDTODEVICE, but works with an interface ID. It
allows binding a socket to a given interface, sending on the socket bypasses
normal routing.

This is for DHCP. It gives us a way to broadcast without any address/route
configured on the interface, without resorting to implementing raw sockets and
constructing IP/UDP packets in userspace.