Rewrite image build process
We now build a "manifest" describing what goes into the filesystem. This gets
written out to a file with a checksum of each file. The FS image build depends
on that file. The effect of this is that the image won't be rebuilt if the
image contents haven't changed.

This was not previously possible to do with just SCons dependencies because we
also have symlinks to set up in the filesystem, and we can't track changes to
those just through SCons.

This is also preparation for two follow up bits of work:
 1. Building hard disk images. I want the image we build to be persistent to
    allow users to keep data in the image. This means we'll need to update the
    image to match what's in the manifest (including removing things as
    needed), while keeping user data. Manifests will let us do this, as we can
    write a manifest of what's currently in the image, and diff them to get
    the changes we need to make.
 2. A rudimentary package system for 3rd party software ports. I don't want to
    have to integrate everything I port into SCons, plus some software will be
    too big to always build. So I'm going to build binary packages. Packages
    will have manifests, which will be used to reduce overhead of tracking
    changes to their files and copying their contents into the image.