Use a persistent disk image for the QEMU target in which user data is preserved
After being built for the first time, every time the image needs to be
updated, we calculate the difference between what we last updated the image
with and the current built system, and apply the difference. This is done
using the manifest. A copy of the manifest is written into the image at the
end of each update, which we read back to compare against at the start of an
update. The update will only change what differs in the manifests: if files
have not changed, we do not bother writing them.

This allows users to persist data in the disk image across runs, while still
updating the image with the built system. This is not yet useful since we
don't have ext2 write support, but this is hopefully Coming Soon(TM).

This will also give a build time improvement once we have large pre-built
packages of 3rd party software to add to the image. Since we only update files
that have changed, we do not need to spend time copying packages into the
image every time we run.