Initial package system
This adds a system for handling binary packages. This will be used for 3rd
party software ports going forward, since it is impractical to integrate all
software we port into SCons and build everything from source all the time.

This is only a build-time system. One day we might have a runtime package
system, but I think that day is pretty far away.

This system is implemented in such a way that overhead of tracking when
packages need to be updated is minimised, plus contents of packages are not
declared to SCons to remove overhead of having it track/checksum files in
them. It uses the manifest system to get the contents of packages into
images (and update them in the incremental image). See comments for full
details.

Package build scripts and binaries are in a submodule for now. This may need
to be changed if we have trouble with GitHub file size limits.

Future changes will make it possible to build things within the main repo
against packages. This will enable freetype to be moved out to a package, for
example. We will also need to be able to build a toolchain sysroot with
packages, so that packages can be built against other packages.