New sysroot build procedure
This replaces the previous sysroot update mechanism that was done every time
SCons was invoked. Now we use a manifest to define the contents of the
sysroot, and it is only updated when explicitly asked for by a build target.

This is a precursor to adding packages to the sysroot so that we can build
against libraries that come from packages.