Fix a bunch of issues with sysroot setup
Since the sysroot was changed to not be always generated, there was an issue
that happened to still work fine on my system until I cleared out some stuff:
we weren't symlinking the core runtime libraries (compiler-rt, crt*.o) in the
sysroot unless explicitly building the sysroot, so new builds without
generating a sysroot would fail.

So, first, we have to always create the compiler-rt link (this lives in the
clang generic install directory, not the sysroot) before building.

Second, don't depend on the sysroot generated by the 'sysroot' target for the
main build. Instead, point the sysroot to our build directory.

That change exposed another issue with packages: we couldn't correctly link
to package libraries that indirectly depend on another without the sysroot.
Need to pass -rpath-link to the linker with the right path to find these.