Replace kernel PRNG with MT19937-64, add pseudo_random device
The previous PRNG we were using in the kernel was *awful*, repeated its
sequence within a few hundred calls. Replace this with MT19937-64.

OpenSSL needs a system random source, so expose the kernel PRNG to userspace
via the /virtual/pseudo_random device, linked from /dev/urandom in the FS.
This is not good for cryptographic use; what we really need is something like
Linux has which sources entropy from system activity and such. However, we're
not a production OS, so this is "ok" for now.